refactor: routers folder - rename auth_utility folder, move utility.py into it and rename
