+++ backend/server/routers/planner.py
@@ -10,11 +10,11 @@ from algorithms.autoplanning import autoplan
 from algorithms.transcript import parse_transcript
 from algorithms.validate_term_planner import validate_terms
 from fastapi import APIRouter, HTTPException, Security, UploadFile
-from server.routers.auth_utility.sessions.middleware import HTTPBearerToUserID
-from server.routers.auth_utility.user import get_setup_user, set_user
+from server.routers.utility.sessions.middleware import HTTPBearerToUserID
+from server.routers.utility.user import get_setup_user, set_user
 from server.routers.model import (CourseCode, PlannedToTerm, PlannerData, ProgramTime, Storage, UnPlannedToTerm,
                                   ValidCoursesState, ValidPlannerData, markMap)
-from server.routers.utility import get_course_details, get_course_object
+from server.routers.utility.common import get_course_details, get_course_object
 
 MIN_COMPLETED_COURSE_UOC = 6
