+++ backend/server/routers/programs.py
@@ -9,7 +9,7 @@ from server.db.mongo.conn import programsCOL
 from server.manual_fixes import apply_manual_fixes
 from server.routers.model import (CourseCodes, Courses, CoursesPathDict, Graph, Programs, Structure, StructureContainer,
                                   StructureDict)
-from server.routers.utility import add_program_code_details_to_structure, add_specialisations_to_structure, convert_adj_list_to_edge_list, get_core_courses, get_gen_eds, get_incoming_edges, get_program_structure, prune_edges
+from server.routers.utility.common import add_program_code_details_to_structure, add_specialisations_to_structure, convert_adj_list_to_edge_list, get_core_courses, get_gen_eds, get_incoming_edges, get_program_structure, prune_edges
 
 router = APIRouter(
     prefix="/programs",
