+++ backend/server/routers/utility/sessions/errors.py
